chore: refactor window tracker helper logic
