+++ apps/desktop/src-tauri/src/helpers/mod.rs
@@ -1,0 +1,1 @@
+pub mod app;
